refactor logs page and add nav link
