+++ webpack/logs/__tests__/index_test.tsx
@@ -1,23 +1,15 @@
-const mockDevice = {
-  updateConfig: jest.fn(() => { return Promise.resolve(); }),
-};
-jest.mock("../../device", () => ({
-  getDevice: () => (mockDevice)
-}));
-
 jest.mock("react-redux", () => ({
   connect: jest.fn()
 }));
 
 import * as React from "react";
 import { mount } from "enzyme";
-import { Logs, LogsFilterMenu, LogsSettingsMenu } from "../index";
+import { Logs } from "../index";
 import { ToolTips } from "../../constants";
 import { TaggedLog, SpecialStatus } from "../../resources/tagged_resources";
 import { Log } from "../../interfaces";
 import { generateUuid } from "../../resources/util";
 import { bot } from "../../__test_support__/fake_state/bot";
-import { ConfigurationName } from "farmbot";
 
 describe("<Logs />", () => {
   function fakeLogs(): TaggedLog[] {
@@ -79,48 +71,4 @@ describe("<Logs />", () => {
     expect(wrapper.text()).toContain("Unknown");
     expect(wrapper.text()).toContain("0, 1, 2");
   });
-});
-
-describe("<LogsFilterMenu />", () => {
-  const fakeState = {
-    autoscroll: true, success: true, busy: true, warn: true,
-    error: true, info: true, fun: true, debug: true
-  };
-  it("renders", () => {
-    const wrapper = mount(
-      <LogsFilterMenu toggle={jest.fn()} state={fakeState} />);
-    ["success", "busy", "warn", "error", "info", "fun", "debug"]
-      .map(string =>
-        expect(wrapper.text().toLowerCase()).toContain(string.toLowerCase()));
-    expect(wrapper.text()).not.toContain("autscroll");
-  });
-
-  it("filters logs", () => {
-    const toggle = jest.fn();
-    const wrapper = mount(
-      <LogsFilterMenu toggle={(x) => () => toggle(x)} state={fakeState} />);
-    wrapper.find("button").first().simulate("click");
-    expect(toggle).toHaveBeenCalledWith("success");
-  });
-});
-
-describe("<LogsSettingsMenu />", () => {
-  it("renders", () => {
-    const wrapper = mount(<LogsSettingsMenu {...bot} />);
-    ["begin", "steps", "complete"].map(string =>
-      expect(wrapper.text().toLowerCase()).toContain(string));
-  });
-
-  function testSettingToggle(setting: ConfigurationName, position: number) {
-    it("toggles setting", () => {
-      bot.hardware.configuration[setting] = false;
-      const wrapper = mount(<LogsSettingsMenu {...bot} />);
-      wrapper.find("button").at(position).simulate("click");
-      expect(mockDevice.updateConfig)
-        .toHaveBeenCalledWith({ [setting]: true });
-    });
-  }
-  testSettingToggle("sequence_init_log", 0);
-  testSettingToggle("sequence_body_log", 1);
-  testSettingToggle("sequence_complete_log", 2);
 });
